Add regridding function with tests.
This implements the third and simplest option, fixing #13. Basic tests are
included, and I've visually confirmed the result makes sense. There seems to be
some odd behavior when using it to upsample diffeomorphisms in the notebook I'm
working on but that is almost surely due to some conceptual problems in the
model.

A simple superresolution atlas with mnist should be the first example notebook
added that uses this.