Add data and utils submodules, affine atlas code
A wrapper for tqdm is available in the new utils module, which detects
whether the program is running inside jupyter or on the command line,
and imports the appropriate tqdm variant.

The data submodule includes the H5Dataset class, a unified interface to
image datasets. The command line interface for this class includes a
single subcommand which downscales a dataset.

The affine module currently includes a command line interface with two
subcommands: one for creating an affine atlas, and another for using the
atlas output to standardize a dataset by applying the inverse
transforms.

These changes required refactoring the existing affine submodule (and
the new data submodule) into __init__.py and __main__.py. This was
necessary for them to be used via e.g. `python -m lagomorph.affine`
without warnings and undefined behavior.

The next step is to add the LDDMM atlas building code and related CLI.