Support multiple tensors in H5Dataset
This allows the 'key' argument to be a tuple, in which case multiple
datasets are returned as torch tensors. The dtype and force_dim
arguments apply only to the first element of those tuples, representing
the assumption that those are images and the other datasets are labels.
These arguments might be needless complication, and I will consider
removing them in the future. For now, they make interacting with older
H5 files that do not specify the channel dimension a little easier.