Add _initialize_compute to Tool
This unifies our approach to parallelism. Any command line tool will
parse the command line and MPI environment in a uniform way. On Summit,
this corresponds to calling jsrun with `jsrun -n<N> -a6 -g6` just as is
expected by horovod. We need MPI>=3 in order to find local rank using
the method implemented here. In the future, we need a fallback to remove
this requirement (see Issue #17).