Add indexed and cached dataloader
This is a large commit that gets rid of the `return_indices` stuff, replacing
that with IndexedDataset which can transform any Dataset into one that returns
indices as well as content. In addition, I added the PreCachedDataset,
LazyCachedDataset, and CachedDataLoader classes, which save data elements or
minibatches to disk using `torch.save`. On Summit, using NVMe this drastically
improved the loading of an HDF5 dataset (OASIS3) from about 90 seconds to under
2 seconds from one epoch to the next.

Note that there is still a lingering bug that's actually a pytorch bug.
`num_workers` should be set to zero on multiGPU machines for now:
https://github.com/pytorch/pytorch/issues/1355