Switch AT_ASSERTM to TORCH_CHECK. Require torch>=1.0
I alias AT_CHECK and TORCH_CHECK in cases where neither are available.
Since PyTorch 1.0 was the first stable release, it makes sense to try to
keep compatibility. Once we have CI set up, we should maintain an env
with 1.0 in it for the foreseeable future.

This change does not affect calling code and is considered a
non-breaking change.

This commit fixes #18.